requirements.txt
streamlit==1.37.1
bokeh==2.4.3
streamlit-bokeh-events==0.1.2
deepgram-sdk==3.5.2
python-dotenv==1.0.0
requests==2.32.3
numpy==1.26.4
groq==0.11.0
